+++ src/helpers/events.test.ts
@@ -1,87 +1,12 @@
-import { beforeEach, it, expect, describe, vi } from 'vitest'
-import { DateInformation, EventDate, ZoneList } from './date-time.types'
+import { it, expect, describe, vi } from 'vitest'
+
+import { DateInformation, EventDate, ZoneList } from './events.types'
+
 import {
-  isValidTimeZone,
-  convertGmtToNumber,
-  getRegionNames,
   getDateInformation,
   getDatesList,
   sortDatesList
-} from './date-time'
-
-describe('isValidTimeZone()', () => {
-  beforeEach(() => {
-    vi.clearAllMocks()
-    vi.resetAllMocks()
-  })
-
-  it('should return true when using a correct time zone', () => {
-    const value = 'Australia/Adelaide'
-    const result = isValidTimeZone(value)
-    expect(result).toBeTruthy()
-  })
-
-  it('should return false when using an incorrect time zone', () => {
-    const value = 'Foo/Bar'
-    const result = isValidTimeZone(value)
-    expect(result).toBeFalsy()
-  })
-
-  it('should return false when Intl is not available', () => {
-    // FIXME: change any types https://vitest.dev/api/vi.html#vi-spyon
-    const spy = vi.spyOn<any, any>(Intl, 'DateTimeFormat')
-    spy.mockImplementation(() => undefined)
-
-    const value = 'Foo/Bar'
-    const result = isValidTimeZone(value)
-    expect(result).toBeFalsy()
-
-    spy.mockRestore()
-  })
-})
-
-describe('convertGmtToNumber()', () => {
-  it('should return an integer on GMT with exact time', () => {
-    const value = 'GMT+1'
-    const result = convertGmtToNumber(value)
-    expect(result).toBe(1)
-  })
-
-  it('should return a decimal on GMT with non exact time ', () => {
-    const value = 'GMT+5:45'
-    const result = convertGmtToNumber(value)
-    expect(result).toBe(5.75)
-  })
-
-  it('should return a decimal on empty GMT', () => {
-    const value = ''
-    const result = convertGmtToNumber(value)
-    expect(result).toBeNull()
-  })
-})
-
-describe('getRegionNames()', () => {
-  it('should return the region name in english', () => {
-    const locale = 'en-US'
-    const countryCode = 'US'
-    const result = getRegionNames(countryCode, locale)
-    expect(result).toBe('United States')
-  })
-
-  it('should return the region name in traditional chinese', () => {
-    const locale = 'zh-Hant'
-    const countryCode = 'US'
-    const result = getRegionNames(countryCode, locale)
-    expect(result).toBe('美國')
-  })
-
-  it('should return the region name with default language', () => {
-    const locale = undefined
-    const countryCode = 'US'
-    const result = getRegionNames(countryCode, locale)
-    expect(result).toBe('United States')
-  })
-})
+} from './events'
 
 describe('getDateInformation()', () => {
   it('should return the date information with 12 time format', () => {
@@ -292,7 +217,7 @@ describe('sortDatesList()', () => {
     expect(result).toMatchObject(expected)
   })
 
-  it('should sort a zone list with null offset', () => {
+  it('should sort a zone list with undefined offset', () => {
     const valueList: EventDate[] = [
       {
         countryCode: 'ES',
@@ -301,7 +226,7 @@ describe('sortDatesList()', () => {
         time: '03:36:42',
         acronym: 'GMT+2',
         gmt: 'GMT+2',
-        offset: null,
+        offset: undefined,
         i18n: {
           region: 'Spain',
           timestamp: '6/10/2023, 03:36:42',
@@ -316,7 +241,7 @@ describe('sortDatesList()', () => {
         time: '21:36:42',
         acronym: 'EDT',
         gmt: 'GMT-4',
-        offset: null,
+        offset: undefined,
         i18n: {
           region: 'United States',
           timestamp: '6/9/2023, 21:36:42',
@@ -334,7 +259,7 @@ describe('sortDatesList()', () => {
         time: '21:36:42',
         acronym: 'EDT',
         gmt: 'GMT-4',
-        offset: null,
+        offset: undefined,
         i18n: {
           region: 'United States',
           timestamp: '6/9/2023, 21:36:42',
@@ -349,7 +274,7 @@ describe('sortDatesList()', () => {
         time: '03:36:42',
         acronym: 'GMT+2',
         gmt: 'GMT+2',
-        offset: null,
+        offset: undefined,
         i18n: {
           region: 'Spain',
           timestamp: '6/10/2023, 03:36:42',
